+++ R/observe_helpers.R
@@ -3,29 +3,30 @@
 #' Function to show a modal dialog, observing each of the help icons in the app.
 #' 
 #' @export
-#' @param input The input object in your shiny app
+#' @param session The session object in your shiny app
 #' @param help_dir A character string of the directory containing your helpfiles
 #' 
 #' @examples
 #' server <- function(input, output, session){
 #' 
 #'   # use anywhere in your server.R script
-#'   observe_helpers(input)
+#'   observe_helpers()
 #'   
 #'   # rest of server.R
 #'   # ...
 #'   # ...
 #' }
 #' 
-observe_helpers <- function(input, help_dir = "helpfiles") {
+observe_helpers <- function(session = shiny::getDefaultReactiveDomain(),
+                            help_dir = "helpfiles") {
   
   shiny::observeEvent(
     
-    eventExpr = input$`shinyhelper-modal_params`,
+    eventExpr = session$input$`shinyhelper-modal_params`,
     
     handlerExpr = {
       
-      params <- input$`shinyhelper-modal_params`
+      params <- session$input$`shinyhelper-modal_params`
       
       type <- params$type
       size <- params$size